Reverted declarations, iterators, etc. to pre-c++11

+++ branches/seabreeze_pre_cpp11/include/api/seabreezeapi/SpectrometerFeatureAdapter.h
@@ -64,7 +64,7 @@ namespace seabreeze {
 			int getOpticalDarkPixelIndices(int *errorCode, int *indices, int length);
 			int getActivePixelCount(int *errorCode);
 			int getActivePixelIndices(int *errorCode, int *indices, int length);
-			unsigned int getNumberOfPixels(int *errorCode);
+			unsigned short getNumberOfPixels(int *errorCode);
             void setIntegrationTimeMicros(int *errorCode,
                     unsigned long integrationTimeMicros);
             long getMinimumIntegrationTimeMicros(int *errorCode);
